Upgrade Hailstorm CLI to rubocop-89.

+++ hailstorm-cli/features/step_definitions/command_line_steps.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 Before('@aws') do
-  $delete_vpc_once ||= false
-  if !$delete_vpc_once
+  $delete_vpc_once ||= false # rubocop: disable Style/GlobalVars
+  unless $delete_vpc_once # rubocop: disable Style/GlobalVars
     delete_vpc_if_exists('hailstorm_cli_integration', 'us-east-2')
-    $delete_vpc_once = true
+    $delete_vpc_once = true # rubocop: disable Style/GlobalVars
   end
 end
 
@@ -12,19 +14,19 @@ Given(/^the "([^"]+)" command line processor is ready$/) do |app_name|
   require 'hailstorm/controller/cli'
   Hailstorm::Initializer.create_middleware(app_name, File.join(tmp_path, app_name, 'config', 'boot.rb'))
   @cli = Hailstorm::Controller::Cli.new(Hailstorm.application)
-  @app_name = app_name
+  @current_project = app_name
 end
 
 
 Given(/^I have a log file on the local filesystem$/) do
-  expect(Dir["#{data_path}/*.jtl"].count).to be > 0
+  expect(Dir["#{data_path}/*.jtl"].count).to be_positive
 end
 
 
 When(/^I import results(?: from '([^']+)'|)$/) do |paths|
-  write_config(@monitor_active) if config_changed?
+  write_config(monitor_active: @monitor_active) if config_changed?
   if paths
-    cmd_args = paths.split(/\s*,\s*/).map { |p| File.join(data_path, p)}.join(' ')
+    cmd_args = paths.split(/\s*,\s*/).map { |p| File.join(data_path, p) }.join(' ')
     @cli.process_cmd_line("results import #{cmd_args}")
   else
     @cli.process_cmd_line('results import')
@@ -45,7 +47,7 @@ Then(/^(\d+) log files? should be imported$/) do |expected_logs_count|
   require 'hailstorm/model/client_stat'
   expect(Hailstorm::Model::ClientStat
              .joins(execution_cycle: :project)
-             .where(projects: {id: @cli.current_project.id}).count).to be == expected_logs_count.to_i
+             .where(projects: { id: @cli.current_project.id }).count).to be == expected_logs_count.to_i
 end
 
 
@@ -56,8 +58,8 @@ end
 
 When(/^(?:I |)(capture the output of|execute) the command "([^"]*)"$/) do |capture_execute, command|
   if capture_execute =~ /capture/
-    @last_cmd_out = capture_cmd_output(:stdout) { @cli.process_cmd_line(command) }
-    puts @last_cmd_out
+    @last_cmd_out = capture_cmd_output { @cli.process_cmd_line(command) }
+    log(@last_cmd_out)
   else
     @cli.process_cmd_line(command)
   end
@@ -75,7 +77,7 @@ end
 
 
 And(/^finalize the configuration$/) do
-  write_config(@monitor_active)
+  write_config(monitor_active: @monitor_active)
 end
 
 
@@ -85,7 +87,7 @@ end
 
 
 When(/^(?:I |)stop load generation with '(.+?)'$/) do |wait|
-  @cli.process_cmd_line('stop wait')
+  @cli.process_cmd_line(wait.blank? ? 'stop' : 'stop wait')
 end
 
 
@@ -107,12 +109,10 @@ end
 
 When(/^(?:I |)configure JMeter with following properties$/) do |table|
   [
-      'hailstorm-site-basic.jmx'
+    'hailstorm-site-basic.jmx'
   ].each do |test_file|
     file_path = File.join(tmp_path, current_project, Hailstorm.app_dir, test_file)
-    unless File.exist? file_path
-      FileUtils.cp(File.join(data_path, test_file), file_path)
-    end
+    FileUtils.cp(File.join(data_path, test_file), file_path) unless File.exist?(file_path)
   end
   jmeter_properties(table.hashes)
 end
@@ -125,44 +125,51 @@ end
 
 When(/^(?:I |)configure target monitoring$/) do
   identity_file = File.join(tmp_path, current_project, Hailstorm.config_dir, 'all_purpose.pem')
-  unless File.exists?(identity_file)
+  unless File.exist?(identity_file)
     FileUtils.cp(File.join(data_path, 'all_purpose.pem'),
                  File.join(tmp_path, current_project, Hailstorm.config_dir))
-    File.chmod(0400, identity_file)
+    File.chmod(0o400, identity_file)
   end
 end
 
 
 Then(/^(\d+) (active |)load agents? should exist$/) do |expected_load_agent_count, active|
-  if active.blank?
-    Hailstorm::Model::LoadAgent.count.should == expected_load_agent_count.to_i
-  else
-    Hailstorm::Model::LoadAgent.active.count.should == expected_load_agent_count.to_i
-  end
+  project = Hailstorm::Model::Project.find_by_project_code(@current_project)
+  conditions = { projects: { id: project.id } }
+  conditions[:load_agents] = { active: true } unless active.blank?
+  query = Hailstorm::Model::LoadAgent.joins(jmeter_plan: :project).where(conditions)
+  expect(query.count).to be == expected_load_agent_count.to_i
 end
 
 
 Then(/^(\d+) Jmeter instances? should be running$/) do |expected_pid_count|
-  Hailstorm::Model::LoadAgent.active.where('jmeter_pid IS NOT NULL').count.should == expected_pid_count.to_i
+  project = Hailstorm::Model::Project.find_by_project_code(@current_project)
+  query = Hailstorm::Model::LoadAgent
+          .active
+          .joins(jmeter_plan: :project)
+          .where(projects: { id: project.id })
+          .where('load_agents.jmeter_pid IS NOT NULL')
+
+  expect(query.count).to be == expected_pid_count.to_i
 end
 
 
-When /^(?:I |)wait for (\d+) seconds$/ do |wait_seconds|
+When(/^(?:I |)wait for (\d+) seconds$/) do |wait_seconds|
   sleep(wait_seconds.to_i)
 end
 
 
-When /^(\d+) (total|reportable) execution cycles? should exist$/ do |expected_count, total|
-  if total.to_sym == :reportable
-    Hailstorm::Model::ExecutionCycle.where(:status => 'stopped').count.should == expected_count.to_i
-  else
-    Hailstorm::Model::ExecutionCycle.count.should == expected_count.to_i
-  end
+When(/^(\d+) (total|reportable) execution cycles? should exist$/) do |expected_count, total|
+  project = Hailstorm::Model::Project.find_by_project_code(@current_project)
+  conditions = { projects: { id: project.id } }
+  conditions[:execution_cycles] = { status: 'stopped' } if total.to_sym == :reportable
+  query = Hailstorm::Model::ExecutionCycle.joins(:project).where(conditions)
+  expect(query.count).to be == expected_count.to_i
 end
 
 
-Then /^a report file should be created$/ do
-  expect(Dir[File.join(tmp_path, current_project, Hailstorm.reports_dir, '*.docx')].count).to be > 0
+Then(/^a report file should be created$/) do
+  expect(Dir[File.join(tmp_path, current_project, Hailstorm.reports_dir, '*.docx')].count).to be_positive
 end
